Move CellAttachedDataSerializer to internal namespace
